Refactor and clean experiments across all languages
- C++: remove dead/commented code, fix blas_lapack syntax error, delete
  all-commented eigen_test.cpp, remove O(2^N) Fibonacci template, fix
  float/double mismatch in nn, remove Python pseudocode from nn.h,
  replace hardcoded /home/rostam path in fs.cpp, remove debug cout
  statements from nonhomo, fix goes_to comment
- Python: flatten 5+ levels of nesting in FindHomomorphisms, fix
  deprecated reindex_axis() in helper.py, remove dead commented blocks,
  complete CannyEdgeDetection stub, add TF 1.x API warnings
- Rust: remove unused area() function and dead variables in rust_gtea
- Haskell: remove unused imports in safediv.hs, remove stub imports/
  functions in graph-coloring-fgl
- Clojure: rename hello_wolrd.clj -> hello_world.clj
- Julia: replace TODO with actual docs link

+++ cpp/nn/nn.h
@@ -17,42 +17,10 @@ class nn {
 public:
     nn(int input_nodes, int hidden_nodes, int output_nodes, double learning_rate);
 
-    static float sigmoid(float x);
+    static double sigmoid(double x);
 };
 
 
 
 
 #endif //NN_NN_H
-
-//
-//self.weights_input_hidden = np.random.normal(0.0, pow(self.hidden, -0.5), (self.hidden, self.input))
-//self.weights_hidden_output = np.random.normal(0.0, pow(self.output, -0.5), (self.output, self.hidden))
-//
-//self.activate_function = lambda x: scs.expit(x)
-//pass
-//
-//        def train(self, inputs_list, targets_list):
-//inputs = np.array(inputs_list, ndmin=2).T
-//targets = np.array(targets_list, ndmin=2).T
-//
-//hidden_inputs = np.dot(self.weights_input_hidden, inputs)
-//hidden_outputs = self.activate_function(hidden_inputs)
-//
-//final_inputs = np.dot(self.weights_hidden_output, hidden_outputs)
-//final_outputs = self.activate_function(final_inputs)
-//
-//output_errors = targets - final_outputs
-//hidden_errors = np.dot(self.weights_hidden_output.T, output_errors)
-//
-//self.weights_hidden_output += self.lr * np.dot((output_errors*final_outputs*(1.0 - final_outputs)), np.transpose(hidden_outputs))
-//self.weights_input_hidden += self.lr * np.dot((hidden_errors*hidden_outputs*(1.0 - hidden_outputs)), np.transpose(inputs))
-//pass
-//
-//        def query(self, inputs_list):
-//inputs = np.array(inputs_list, ndmin=2).T
-//hidden_inputs = np.dot(self.weights_input_hidden, inputs)
-//hidden_outputs = self.activate_function(hidden_inputs)
-//final_inputs = np.dot(self.weights_hidden_output, hidden_outputs)
-//final_outputs = self.activate_function(final_inputs)
-//return final_outputs
